# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "tpms back-end Api docs",
        "title": "TPMS-BE Api",
        "contact": {},
        "license": {
            "name": "TMPS"
        },
        "version": "1.0.1"
    },
    "host": "https://tpms-fdwva.ondigitalocean.app/tpms-be2",
    "basePath": "/api/v1",
    "paths": {
        "/dog": {
            "post": {
                "description": "Register new dog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dog"
                ],
                "summary": "Register new dog",
                "parameters": [
                    {
                        "description": "dog",
                        "name": "dog",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.DogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DogResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates existing dog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dog"
                ],
                "summary": "Updates dog",
                "parameters": [
                    {
                        "description": "dog",
                        "name": "dog",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.DogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DogResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dog/:id": {
            "get": {
                "description": "Get dog given its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dog"
                ],
                "summary": "Get dog given its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dog ID",
                        "name": "dog",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DogResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dog/found": {
            "patch": {
                "description": "Reunite dog with owner. Making him its only host and removing other hosts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dog"
                ],
                "summary": "Reunite dog with owner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dog ID",
                        "name": "dogID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dog owner ID",
                        "name": "ownerID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dog host ID",
                        "name": "hostID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DogResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dog/missing": {
            "get": {
                "description": "If no argument is given it returns all missing dogs. If user location and a search radius is sent, then it returns all missing dogs within that radius.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dog"
                ],
                "summary": "Brings list of missing dogs",
                "parameters": [
                    {
                        "type": "number",
                        "description": "user latitude",
                        "name": "userLatitude",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "user longitude",
                        "name": "userLongitude",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "radio to look for dogs",
                        "name": "radius",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DogResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.DogRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "coatColor": {
                    "type": "string"
                },
                "coatLength": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imgUrl": {
                    "type": "string"
                },
                "imgs": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "isLost": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                }
            }
        },
        "model.DogResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "coatColor": {
                    "type": "string"
                },
                "coatLength": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "img": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "imgUrl": {
                    "type": "string"
                },
                "isLost": {
                    "type": "boolean"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                }
            }
        },
...
    }
}